requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
jsonschema==3.2.0
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyrsistent==0.17.3
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
